Add support for sparse crates.io index
